% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openBeginnerExample.R
\name{openBeginnerExampleApp}
\alias{openBeginnerExampleApp}
\title{Open Beginner App}
\usage{
openBeginnerExampleApp()
}
\description{
This function opens the beginner example app for editing.
}
\examples{
openBeginnerExampleApp()
}
