% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openBeginnerIntermediateExampleApp.R
\name{openBeginnerIntermediateExampleApp}
\alias{openBeginnerIntermediateExampleApp}
\title{Open Beginner-Intermediate App}
\usage{
openBeginnerIntermediateExampleApp()
}
\description{
This function opens the Beginner-Intermediate example app for editing.
}
\examples{
openBeginnerIntermediateExampleApp()
}
