% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/openMeetupParticipantGroupingApp.R
\name{openMeetupParticipantGroupingApp}
\alias{openMeetupParticipantGroupingApp}
\title{Open MeetupParticipantGrouping App}
\usage{
openMeetupParticipantGroupingApp()
}
\description{
This function opens the MeetupParticipantGrouping example app for editing.
}
\examples{
openMeetupParticipantGroupingExampleApp()
}
