% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBeginnerExample.R
\name{runBeginnerExample}
\alias{runBeginnerExample}
\title{Run Beginner Example Shiny App}
\usage{
runBeginnerExample()
}
\description{
This function runs the beginner example app.
}
\examples{
runBeginnerExample()
}
