% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runBeginnerIntermediateExample.R
\name{runBeginnerIntermediateExample}
\alias{runBeginnerIntermediateExample}
\title{Run Beginner-Intermediate Example Shiny App}
\usage{
runBeginnerIntermediateExample()
}
\description{
This function runs the Beginner-Intermediate example app.
}
\examples{
runBeginnerIntermediateExample()
}
