% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runMeetupParticipantGroupingExample.R
\name{runMeetupParticipantGroupingExample}
\alias{runMeetupParticipantGroupingExample}
\title{Run MeetupParticipantGrouping Example Shiny App}
\usage{
runMeetupParticipantGroupingExample()
}
\description{
This function runs the MeetupParticipantGrouping example app.
}
\examples{
runMeetupParticipantGroupingExample()
}
